+++ one/__init__.py
@@ -1,2 +1,2 @@
-"""The Open Neurophysiology Environment (ONE) API"""
-__version__ = '2.1.0'
+"""The Open Neurophysiology Environment (ONE) API."""
+__version__ = '2.2.0'